requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-mongoengine==1.0.0
Flask-WTF==1.2.1
graphviz==0.20.3
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.28.2
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
pydot==3.0.1
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
waitress==3.0.0
Werkzeug==2.0.3
WTForms==3.1.2
